Add GitHub Actions workflow to build releases
Will automatically build all commits that are pushed to the project and
produce a build artifact. If a commit is tagged, it will automatically
create a release based on the tag name and attach the build to it.

A simple `tag-release` script has been added to easily create tags with
the same format that is currently used ("1.0.0r<num commits>").