Fix crash when dealing with non-ASCII paths
When running with a non-ASCII path (set via Windows username, the
`-path` arg, or the `path=..` setting in `chimera.ini`), Chimera would
crash, sometimes creating 2 new folders with mangled names.

An example of a problematic path is "ÁÉÚÝÍÓŠ".

This commit fixes this by setting the locale of the dll to match the
system's configured locale.

Doing this fixes the cases of the Windows username and the `-path` arg,
but there can still be issues when setting the path via `chimera.ini`.
Because the ini parser isn't encoding aware, a `chimera.ini` encoded
using UTF-8 will cause a mangled path to be set when non-ASCII
characters are used in it.

The fix for this is to document that `chimera.ini` must be encoded using
the default 8-bit system encoding (sometimes called "ANSI") in the file.

Many thanks to Adam Šliperski for discovering this issue and working
with me to fix it.