Require that chimera.ini be encoded using UTF-8
Values read from the ini file are converted from UTF-8 into the system's
default 8-bit encoding ("ANSI"). If decoding the bytes as UTF-8 fails or
any of the characters can't be converted to ANSI then an error message
will be displayed and Chimera will exit.

Since most text editors now default to saving as UTF-8 this should just
work in most cases. If not, an error message will be shown instead of
just using the mangled value.