Fix signature error messages not showing on startup
The `show_error_box` function only shows a message box if it detects
Chimera is running as a client. This detection is done by checking for
all the client signatures.

In order to show an error message when not all client signatures were
found, an optional `is_client` paramater was added to override the
signature-based check (which would always return false in this case).

A bug was also fixed where if the `show_error_box` function was called
before the `chimera.ini` file was read, Chimera would segfault.