Speed up processing of chat messages
Will now only attempt to remove IP addresses if the message is long
enough to contain one (7+ characters). This greatly improves the
framerate on servers that send 20+ blank messages per tick in order to
display timers and other information on the client.

The regex that replaces IP addresses was also simplified, slightly
improving performance for chat messages 7+ characters long.