Fix bug in drawing text with escaped carets
This commit fixes an issue where carets immediately after an escaped
caret (`^^`) wouldn't be properly detected as starting a color code (or
another escaped caret).

For example, the sequence `^^^1a` was being rendered as `^^1a`. With
this commit applied, it is correctly rendered as a `^` followed by a red
`a`.