More Interpolation Stuff
Been sitting on these for a while. Basically just addresses the scenarios where the interpolation buffers should be cleared in a more robust way.

The buffers will now clear on any event where the game state drastically changes (revert checkpoint, core load, map reset, loading a new map etc) instead of simply when the tick counter changes to be less than what it was. This was specifically to target the edge case scenario of reverting or core loading into the future (ie the tick counter is larger after the revert).