Fix contrail leading parent object/projectile
When interpolation is enabled, the parent object of a contrail trails the front of the contrail, most obvious on slow moving projectiles like the c gun projectile or banshee wingtips. This "fix" rolls back the parent objects position to the previous tick value for the duration of the contrail update function so that contrails trail the object. This matches the appearance as seen in MCC.