Fix specular point lights
The stock game relies on legacy shader model 1.x texture sampling behaviour to handle both point and spot light specular highlights with a single shader (spotlight sets a cube map, whilst point light sets a 3D texture to the same texture sampler). This texture/sampler mismatch is undefined behaviour in shader model 2 and may or may not work depending on driver implementation. The fixed H1A HLSL code cannot be compiled against ps_1_4. Fortunately, the game sets the vector normalization bitmap in samplers 2 and 3. We can use one of these duplicated samplers to be the sampler for the distance attenuation 3D texture whilst leaving the flashlight cube map on sampler 1.  We can then toggle between samplers using a shader constant if we are drawing a spotlight or point light specular.

It goes without saying that this relies on both patches to the draw function and shader code.